Create Entry objects as entries on priority queue
Added Entry object as entries for the priority instead of lists. Through
the code we would reference elements in the list with their index. This
wasn't optimal we want to reference those elements with a dot notation.

However, we wanted the 'pointer' functionality from the lists, and that
is still the the same with this implementation.